Fix missing JavaSourceRootPropertiesEntity when adding new files to modules
When a new file was added and processed by BazelFileEventListener, the
addToModule() function created a SourceRootEntity but never attached a
JavaSourceRootPropertiesEntity with the packagePrefix. This caused:
1. No code intelligence for the newly added file until reopening
2. Alt+Enter import resolution failing for other files importing from it

The fix:
- Consolidate duplicate addToModule() into a single canonical version that
  creates JavaSourceRootPropertiesEntity with the correct packagePrefix
  (resolved from the module's PackageMarkerEntity for the parent directory)
- Remove the dead AssignFileToModuleListener class (not registered in
  plugin.xml since upstream moved to BazelFileEventListener)
- Port downstream features (UnsyncedTargetUpdater partial sync, testlib
  stripping, isTestModule detection) into BazelFileEventListener
- Route external file submissions through BazelFileEventListener's queue
