+++ plugin-bazel/src/main/kotlin/org/jetbrains/bazel/workspace/fileEvents/SimplifiedFileEvent.kt
@@ -78,6 +78,10 @@ internal sealed class SimplifiedFileEvent private constructor(
   class Create(originalEvent: VFileCreateEvent)
     : SimplifiedFileEvent(fileAdded = originalEvent.path, newVirtualFile = originalEvent.file)
 
+  /** Event for externally submitted files (e.g., generated by Bazel build) that need module assignment. */
+  class ExternalCreate(file: VirtualFile)
+    : SimplifiedFileEvent(fileAdded = file.path, newVirtualFile = file)
+
   class CreateDirectory(originalEvent: VFileCreateEvent) :
     SimplifiedFileEvent(fileAdded = originalEvent.path, newVirtualFile = originalEvent.file)
